refactor(setup): rename test directory
