+++ tests/test_main.py
@@ -3,14 +3,12 @@
 import pytest
 from git_author_rewrite.main import (
     git,
-    display_messages,
     main,
     is_email_format_valid,
     create_working_branch,
     sync_working_branch,
 )
 from unittest.mock import MagicMock
-from git_author_rewrite.enums import Status
 
 TEST_BRANCH_NAME = 'author-rewrite'
 
@@ -52,6 +50,7 @@ def test_sync_working_branch_not_on_active_branch(mock_repo):
     # Check if checkout() was called once 
     rewrite_branch.checkout.assert_called_once()
 
+
 def test_sync_working_branch_on_active_branch(mock_repo):
     """Test syncing the working branch with the default branch."""  
     default_branch = mock_repo.heads['main']
@@ -64,55 +63,6 @@ def test_sync_working_branch_on_active_branch(mock_repo):
 
     # Check if checkout() was called once 
     rewrite_branch.checkout.assert_called_once()
-
-
-def test_display_messages_exits_with_error_and_prints_to_stderr(capfd):
-    """Test that the function prints a message to stderr and exits with status code 2."""
-    message = "error"
-
-    with pytest.raises(SystemExit) as excinfo:
-        display_messages(message, status=Status.ERROR)
-
-    # Capture stdout and stderr
-    out, err = capfd.readouterr()
-
-    assert excinfo.value.code == 2, 'Expected an exit code of 2 (stderr)'
-    assert message in err, 'Expected a message'
-
-
-def test_display_messages_prints_to_stdout(capfd):
-    """Test that the function prints a message to stdout"""
-    message = "message"
-
-    display_messages(message)
-
-    out, err = capfd.readouterr()
-
-    assert message in out, 'Expected a message'
-
-
-def test_display_messages_exits_and_prints_multiple_messages_to_stderr(capfd):
-    """Test that the function prints multiple messages to stderr and exits."""
-
-    messages = ["error 1", "error 2"]
-    with pytest.raises(SystemExit) as excinfo:
-        display_messages(*messages, status=Status.ERROR)
-
-    out, err = capfd.readouterr()
-
-    assert excinfo.value.code == 2, 'Expected an exit code of 2 (stderr)'
-    assert "\n".join(messages) in err, 'Expected two messages';
-
-
-def test_display_messages_prints_multiple_mesages_to_stdout(capfd):
-    """Test that the function prints multiple messages to stdout."""  
-
-    messages = ["message 1", "message 2"]
-
-    display_messages(*messages)
-
-    out, err = capfd.readouterr()
-    assert "\n".join(messages) in out, 'Expected two messages';
 
 
 def test_is_email_format_valid_invalid_email():
